fix(tests): renaming paths for collection cards
